Rename dots TTS native/ to _vendored/ and drop dead training dir
Rename the vendored tree native/ -> _vendored/ to match the house convention
(cf. higgs_tts/_vendored) and make provenance self-evident. Pure path rename:
all dots_tts.native.* imports become dots_tts._vendored.* (native_adapter is
untouched); VENDORED.md wording refreshed. Also delete the empty native/training/
shell left over from the earlier training-pipeline prune.
