+++ sglang_omni/models/dots_tts/_vendored/models/dots_tts/model.py
@@ -20,8 +20,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 from einops import rearrange
 from loguru import logger
-from sglang_omni.models.dots_tts.native.utils.profiling import measure_inference
-from sglang_omni.models.dots_tts.native.utils.util import get_dtype
+from sglang_omni.models.dots_tts._vendored.utils.profiling import measure_inference
+from sglang_omni.models.dots_tts._vendored.utils.util import get_dtype
 
 
 @dataclass
